# SpringBoot/08. rest-api/challenge-testing/starter/contacts/src/main/resources/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Contact API",
        "description": "An API that can manage contacts",
        "version": "v1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "tags": [
        {
            "name": "Contact Controller",
            "description": "Create and retrieve contacts"
        }
    ],
    "paths": {
        "/contact": {
            "post": {
                "tags": [
                    "Section Three",
                    "Contact Controller"
                ],
                "summary": "Create Contact",
                "description": "Creates a contact from the provided payload",
                "operationId": "createContact",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Contact"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful creation of contact",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Contact"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request: unsuccessful update",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/contact/{id}": {
            "get": {
                "tags": [
                    "Section Two",
                    "Contact Controller"
                ],
                "summary": "Get contact by Id",
                "description": "Returns a contact based on an ID",
                "operationId": "getContact",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful retrieval of contact",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Contact"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Contact doesn't exist",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/contact/all": {
            "get": {
                "tags": [
                    "Contact Controller",
                    "Section One"
                ],
                "summary": "Retrieves contacts",
                "description": "Provides a list of all contacts",
                "operationId": "getContacts",
                "responses": {
                    "200": {
                        "description": "Successful retrieval of contacts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Contact"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Contact": {
                "required": [
                    "name",
                    "phoneNumber"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "phoneNumber": {
                        "type": "string"
                    }
                }
            },
            "ErrorResponse": {
                "type": "object",
                "properties": {
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "message": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}